Revert "texcache-rebase"
This reverts commit f0b480135ab0b5155cb51829ac316b944a6f0943.

+++ src/core/settings.h
@@ -95,7 +95,7 @@ struct Values {
     // Renderer
     bool use_hw_renderer;
     bool use_shader_jit;
-    u16 resolution_factor;
+    float resolution_factor;
     bool use_vsync;
     bool toggle_framelimit;
